Moved BatchProtocols to dedicated module
Pure move, nothing else
