+++ tianshou/policy/imitation/bcq.py
@@ -6,7 +6,8 @@ import torch
 import torch.nn.functional as F
 
 from tianshou.data import Batch, to_torch
-from tianshou.data.batch import BatchProtocol, RolloutBatchProtocol
+from tianshou.data.batch import BatchProtocol
+from tianshou.data.types import RolloutBatchProtocol
 from tianshou.policy import BasePolicy
 from tianshou.utils.net.continuous import VAE
