+++ tianshou/policy/imitation/cql.py
@@ -6,7 +6,7 @@ import torch.nn.functional as F
 from torch.nn.utils import clip_grad_norm_
 
 from tianshou.data import Batch, ReplayBuffer, to_torch
-from tianshou.data.batch import RolloutBatchProtocol
+from tianshou.data.types import RolloutBatchProtocol
 from tianshou.policy import SACPolicy
 from tianshou.utils.net.continuous import ActorProb
