+++ tianshou/policy/imitation/discrete_cql.py
@@ -5,7 +5,7 @@ import torch
 import torch.nn.functional as F
 
 from tianshou.data import to_torch
-from tianshou.data.batch import RolloutBatchProtocol
+from tianshou.data.types import RolloutBatchProtocol
 from tianshou.policy import QRDQNPolicy
