+++ tianshou/policy/imitation/gail.py
@@ -5,7 +5,7 @@ import torch
 import torch.nn.functional as F
 
 from tianshou.data import ReplayBuffer, to_numpy, to_torch
-from tianshou.data.batch import RolloutBatchProtocol
+from tianshou.data.types import RolloutBatchProtocol
 from tianshou.policy import PPOPolicy
 from tianshou.policy.modelfree.pg import TDistParams
 from tianshou.policy.modelfree.ppo import BatchWithLogpOldProtocol
