+++ tianshou/policy/modelbased/psrl.py
@@ -4,9 +4,9 @@ import numpy as np
 import torch
 
 from tianshou.data import Batch
-from tianshou.data.batch import BatchProtocol, RolloutBatchProtocol
+from tianshou.data.batch import BatchProtocol
+from tianshou.data.types import ActBatchProtocol, RolloutBatchProtocol
 from tianshou.policy import BasePolicy
-from tianshou.policy.base import ActBatchProtocol
 
 
 class PSRLModel(object):
