+++ tianshou/policy/modelfree/a2c.py
@@ -6,7 +6,7 @@ import torch.nn.functional as F
 from torch import nn
 
 from tianshou.data import ReplayBuffer, to_torch_as
-from tianshou.data.batch import BatchWithReturnsProtocol, RolloutBatchProtocol
+from tianshou.data.types import BatchWithReturnsProtocol, RolloutBatchProtocol
 from tianshou.policy import PGPolicy
 from tianshou.policy.modelfree.pg import TDistParams
 from tianshou.utils.net.common import ActorCritic
