+++ tianshou/policy/modelfree/c51.py
@@ -4,7 +4,7 @@ import numpy as np
 import torch
 
 from tianshou.data import ReplayBuffer
-from tianshou.data.batch import RolloutBatchProtocol
+from tianshou.data.types import RolloutBatchProtocol
 from tianshou.policy import DQNPolicy
