+++ tianshou/policy/modelfree/dqn.py
@@ -5,13 +5,13 @@ import numpy as np
 import torch
 
 from tianshou.data import Batch, ReplayBuffer, to_numpy, to_torch_as
-from tianshou.data.batch import (
-    BatchProtocol,
+from tianshou.data.batch import BatchProtocol
+from tianshou.data.types import (
     BatchWithReturnsProtocol,
+    ModelOutputBatchProtocol,
     RolloutBatchProtocol,
 )
 from tianshou.policy import BasePolicy
-from tianshou.policy.base import ModelOutputBatchProtocol
 
 
 class DQNPolicy(BasePolicy):
