+++ tianshou/policy/modelfree/ppo.py
@@ -5,7 +5,7 @@ import torch
 from torch import nn
 
 from tianshou.data import ReplayBuffer, to_torch_as
-from tianshou.data.batch import RolloutBatchProtocol
+from tianshou.data.types import RolloutBatchProtocol
 from tianshou.policy import A2CPolicy
 from tianshou.policy.modelfree.a2c import BatchWithAdvantagesProtocol
 from tianshou.policy.modelfree.pg import TDistParams
