+++ tianshou/policy/modelfree/qrdqn.py
@@ -6,7 +6,7 @@ import torch
 import torch.nn.functional as F
 
 from tianshou.data import ReplayBuffer
-from tianshou.data.batch import RolloutBatchProtocol
+from tianshou.data.types import RolloutBatchProtocol
 from tianshou.policy import DQNPolicy
